+++ test/package_api_differ_test_data.dart
@@ -430,6 +430,8 @@ final simpleFieldDeclarationA = FieldDeclaration(
   isExperimental: false,
   isStatic: false,
   relativePath: '',
+  isReadable: true,
+  isWriteable: true,
 );
 final simpleFieldDeclarationB = FieldDeclaration(
   typeName: 'bool',
@@ -439,6 +441,8 @@ final simpleFieldDeclarationB = FieldDeclaration(
   isExperimental: false,
   isStatic: false,
   relativePath: '',
+  isReadable: true,
+  isWriteable: true,
 );
 
 final packageFieldA = PackageApi(