+++ test/test_packages/field_to_property/package_a_with_property/lib/src/class_a.dart
@@ -1,0 +1,6 @@
+class ClassA {
+  String _stringValue;
+  String get stringValue => _stringValue;
+
+  ClassA(String stringValue) : _stringValue = stringValue;
+}